OmniSwitch: locked apps support
allows tagging apps as "locked" which means they will not
be killed on actions kill all and kill other

explicitly killing a locked task will remove it from
the locked list and kill it in one step.

the locked apps are stored as a list
of package names. this list is updated on package changes
like uninstall and will remove uninstalled apps then

also add an option to sort locked apps on top
of the recent list

+++ src/org/omnirom/omniswitch/SwitchManager.java
@@ -231,6 +231,11 @@ public class SwitchManager {
             hide(false);
         }
 
+        if (ad.isLocked()) {
+            // remove from locked
+            toggleLockedApp(ad, ad.isLocked(), false);
+        }
+
         mAm.removeTask(ad.getPersistentTaskId());
         if (DEBUG){
             Log.d(TAG, "kill " + ad.getLabel());
@@ -266,6 +271,9 @@ public class SwitchManager {
         Iterator<TaskDescription> nextTask = mLoadedTasks.iterator();
         while (nextTask.hasNext()) {
             TaskDescription ad = nextTask.next();
+            if (ad.isLocked()) {
+                continue;
+            }
             mAm.removeTask(ad.getPersistentTaskId());
             if (DEBUG){
                 Log.d(TAG, "kill " + ad.getLabel());
@@ -291,6 +299,9 @@ public class SwitchManager {
         nextTask.next();
         while (nextTask.hasNext()) {
             TaskDescription ad = nextTask.next();
+            if (ad.isLocked()) {
+                continue;
+            }
             mAm.removeTask(ad.getPersistentTaskId());
             if (DEBUG){
                 Log.d(TAG, "kill " + ad.getLabel());
@@ -316,6 +327,10 @@ public class SwitchManager {
 
         if (getTasks().size() >= 1){
             TaskDescription ad = getTasks().get(0);
+            if (ad.isLocked()) {
+                // remove from locked
+                toggleLockedApp(ad, ad.isLocked(), false);
+            }
             mAm.removeTask(ad.getPersistentTaskId());
             if (DEBUG){
                 Log.d(TAG, "kill " + ad.getLabel());
@@ -692,6 +707,23 @@ public class SwitchManager {
             // make sure we stay on the correct focus
             // and killing is not changing it - overlay stays open here
             restoreHomeStack();
+            mLayout.refresh();
+        }
+    }
+
+    public void toggleLockedApp(TaskDescription ad, boolean isLockedApp, boolean refresh) {
+        List<String> lockedAppsList = mConfiguration.mLockedAppList;
+        if (DEBUG){
+            Log.d(TAG, "toggleLockedApp " + lockedAppsList);
+        }
+        if (isLockedApp) {
+            Utils.removedFromLockedApps(mContext, ad.getPackageName(), lockedAppsList);
+        } else {
+            Utils.addToLockedApps(mContext, ad.getPackageName(), lockedAppsList);
+        }
+        ad.setLocked(!isLockedApp);
+        ad.setNeedsUpdate(true);
+        if (refresh) {
             mLayout.refresh();
         }
     }
